Moved more of the tests to the AbstractChoiceFieldTest.php
